passando funcao productAmountfilter e estado dos itens filtrados para o APP
